ci: deflake macOS builds with CMake (#9306)
I searched for places where there is an explicit or implicit download,
and wrap them with a retry loop. This motivated me to refactor the
code to install vcpkg.

Incidentally, my previous changes to use GCS for the vcpkg cache may
have increased flakiness: instead of one download for the vcpkg binary
caches it performs N (currently 14). This is fixed in this change too.