doc: improve documentation around Abseil's ABI pinning (#9334)
Our instructions did not highlight why we recommend the magic `sed`
command to tweak Abseil. The command is easy to miss, and deserves a
larger warning. On macOS with homebrew this is even more tricky, as
homebrew does not patch Abseil's headers.

I also updated the instructions around enabling additional libraries
with CMake.