feat(common): truncation support for plain strings in the RPC log
Add `DebugString(std::string, TracingOptions const&)` for adding a plain
string to the RPC log, but with the `truncate_string_field_longer_than`
behavior of `google::protobuf::TextFormat::Printer`.

This will be used in an upcoming change that will log a `resume_token`
from the Spanner streaming read/query protocol, and which is extremely
long.