test(storage): make benchmark more memory efficient (#9358)
The benchmark was allocating a (constant) buffer of random data to
upload on each thread. The default size of this buffer is 64 MiB.
This is a good default, as around 64 MiB the performance plateaus.
We also are starting to run benchmarks with 64 and more threads,
and each thread several buffers (one per transport).  That quickly
pushes the memory consumption into the 100 GiB range.
Better to share the buffer across all these threads.