impl(spanner): fix the handling of PartialResultSet.resume_token (#9359)
The service only adds a `resume_token` when the `PartialResultSet` ends
with a completed row (except when the `PartialResultSet` also completes
the stream, in which case the `resume_token` remains empty). Therefore,
`resume_token` and `chunked_value` are mutually exclusive.

This means the library has to buffer rows completed within a non-token
`PartialResultSet` if it is to maintain resumability. This mainly affects
the algorithm and data used by `PartialResultSetSource::ReadFromStream()`.
For example, the storage of the token moves from `PartialResultSetResume`
to `PartialResultSetSource`, and the latter gains a `deque` of complete,
but undeliverable rows.

Introduce the `StreamingResumabilityBufferSizeOption` to control the size
of the buffer. If this size is exceeded, complete rows will be delivered
to the user, which makes the stream non-resumable until a new `resume_token`
is received.

Correct the `ConnectionImplTest.ReadSuccess` test case, which was checking
that we did the wrong thing. :-)

Add new `PartialResultSetResume` test cases for stream resumption.

Loop within the `PartialResultSetSourceTest.MultipleResponses` test to
validate that different buffering limits do not affect a complete stream.

Fixes #8523.

While we're here, canonicalize the shape of protobuf text in tests, which
includes giving resume tokens meaningful content, and also include some
additional expectations that `BeginTransaction()` is not called.