impl(pubsub): use unary RPCs to refresh leases
So far we have been using the `StreamingPull()` RPC to receive data and
to refresh the message leases. To implement exactly-once delivery we
need to retry the message lease refreshes, and it is better to do this
using a separate unary RPC because (1) it is easier to implement,
(2) it is more consistent with other libraries, and (3) the preferred
approach by the Pub/Sub service folks.