test(storage): benchmark supports ranged reads (#9378)
A lot of analytical workloads use "ranged" reads, i.e., they read a
portion of a larger object. There seems to be performance advantages
when reading up to 2MiB. Likewise, there are measurable performance
differences when the offset is on a 2MiB boundary and/or the read
size is on a 2 MiB boundary.

We would like to measure these effects, and therefore we need new
parameters for the benchmark.

By now, we have 4 of these "quantized" size ranges, I did some
refactoring to avoid code duplication.