test(storage): more tweaks to ranged reads
The benchmark would perform a range read with 50% probability, but some
ranged reads were "full" as the range was larger than the object. I
found it hard to reason about this, so now only the range size controls
whether the read is full or not.

I also rediscovered a quirk of the JSON API, that it is better avoided:
reading the last 0 bytes actually returns all the bytes. Maybe we should
fix this in the client library itself, for now just avoid running into
it.