chore!: updates as Debian 9 (Stretch) is EOL (#9405)
Debian 9 (Stretch) reached EOL on 2022-06-30. Therefore, we have stopped
testing or supporting this distribution. This was the last distribution
we supported that required GCC < 7.3, and/or CMake < 3.10. Starting with
this release we require GCC >= 7.3 and CMake >= 3.10.

This PR updates the documentation, CI builds and fixes some test problems
that only show up with GCC circa 7.3