cleanup(storage): reduce maximum function cognitive complexity
I ran `clang-tidy` with the `readability-function-cognitive-complexity`
warning enabled. It found some functions that were too complex for their
own good, this cleans up all such functions in the storage library.
There are more such functions to clean up in the gRPC plugin, and I
have not even started on tests.