ci: missing export in build.sh
When building with `--cloud=...` we need to export the
`VERBOSE_FLAG` environment variable. In most command-line builds this
is exported via the `docker run --env=...` commands, so it was hard to
detect. To make it easier to detect such problems, I reordered the
`export` commands to match the `docker run` order.