feat(rest): reduce contention in `PooledCurlHandleFactory` (#9457)
Use separate mutexes for each component. Move as much work as possible
outside the critical sections. And finally amortize the cost of
releasing the handles to one thread.

Of these changes the most important seems to be moving the work to
fetch the local IP address outside the critical sections, as this may
involve a DNS lookup. The rest are at least not harmful, and could
help when under pressure.

I added a benchmark.  I ran this to 32,768 simultaneous threads using
the pool. It works and only takes a few milliseconds, but this is not
a number of threads we can safely run in the CI environment.