impl(storage): helper class to generate invocation ids
We will be adding telemetry to GCS requests, where all the retry
attempts for the same application call will share an "invocation id".
This will be included in the `x-goog-api-client` header, under the
`gccl-invocation-id/` field. The design doc calls for a very specific
format for these ids, which this helper class generates.