feat: improve error messages for access token errors
When we fail to generate an access token the errors can be very
confusing.  See for example this discussion:

https://issues.apache.org/jira/browse/ARROW-17069

This change is an attempt at making these errors less confusing. They
could be made event better if we had a specific webpage describing
client library authentication and troubleshooting, but we do not have
such a page at the moment.