test(spanner): additional logging on UpdateDatabaseDdl() failures (#9483)
When we see an UpdateDatabaseDdl() failure in the Spanner samples, fetch
the database DDL, and list all completed and pending operations, so that
we can use the RPC log to summarize the service state.  Then abort the
process so the database will not be dropped.  Part of #4758.

The extra code within the samples includes `//!` comments so that it will
be omitted from the extracted snippets.