ci: make it easier to run on ARM64 machines (#9488)
Make it easier to run some builds on ARM64 machines. The Dockerfiles
download a few pre-built binaries, and those need to be platform
specific.

There is no Spanner emulator for ARM64, in this change we just don't
install it. Some builds do not need it, but I did not take any action
(such as disabling the integration tests) beyond that.