test(spanner): MakeConnection() test will require a service (#9545)
`ClientTest.MakeConnectionOptionalArguments` is really an integration
test in unit-test clothing.  Indeed, in the near future just creating
a database connection will start the process of filling the session
pool.  So, move it from `client_test.cc` to `client_integration_test.cc`.

Also, the purpose of the test is to ensure that the previous overloads
with `ConnectionOptions` and `SessionPoolOptions` are still available,
and not to verify any particular behavior.  So, simplify accordingly.