impl(rest): always discard handles via HandleFactory (#9575)
When a CURL* handle has an error we don't want to return it to the pool.
The error probably indicates a bad server (in the load balancer) or a
bad connection, and it is time to set up a new connection. Previously,
we just discarded the handle, how let the factory do so. In a future
PR we will use this to update counters in the factory, and adjust how
aggressively the pool releases its handles.