impl(generator): only apply service default options in one place
Only call `$service_name$DefaultOptions()` from the `Make*Connection()`
factory, rather that also from the "ConnectionImpl" and "Client" ctors.
This makes it very clear where the defaults are coming from, and avoids
redundant work.  (It also opens a path for the default options to depend
on information only available during `Make*Connection()`.)

This does mean that tests that skip `Make*Connection()` will need to
ensure that the mock connection's `options()` expectation supplies the
default options, but this is a feature.