feat(storage): add debugging headers to `ObjectWriteStream`
Make it easier to troubleshoot problems with resumable uploads (aka
`Client::WriteObject()`) by capturing the headers of the last request.
We do not want to capture the headers for all the requests because there
can be an unbounded number of requests, and therefore all the headers
can consume arbitrary amounts of memory.