impl(generator): do not OCC loop when there is no backoff policy
In the OCC-loop overload of SetIamPolicy(), do not loop on kAborted
when there is no BackoffPolicyOption.  This will not affect production
code as there is a default BackoffPolicyOption in that case.  However,
a mock connection in a test may not have one, but we undoubtedly never
want to loop in that circumstance anyway unless the test has expressly
set a policy.