cleanup(storage): remove unused (and buggy) credentials type (#9614)
The `SelfSignedServiceAccountCredentials` were buggy (the name of the
signing algorithm had a typo `HS256` when it should have been `RS256`).
Since the class was unusable, and in `storage::internal` I think we can
safely remove it without breaking anybody.

I am going to add a (fixed) implementation of self-signed JWTs soon.