feat(generator): support services with location-dependent endpoints
Add generator support for services with location-dependent default
endpoints.  A new `MakeServiceConnection()` overload is generated with
a required `location` parameter, which is used to build the default
`EndpointOption` of "<location>-<normal-service-endpoint>" (for
example, "us-documentai.googleapis.com" in the Cloud Document AI API).

For backwards compatibility in already-released libraries, the old
`MakeServiceConnection()` signature is still available, but it, as
before, requires setting `EndpointOption` and `AuthorityOption` to
get something that works.

Fixes #7976 and #8704.