cleanup: auto format markdown files using "mdformat"
See https://mdformat.readthedocs.io/en/stable/users/style.html for a
description of the formatting style that `mdformat` follows.

`.mdformat.toml` specifies the tweakable formatting parameters, only
one of which we change from the default.  We set `number=true` to
minimize the diffs in the initial formatting, but it is probably best
to revert to the default `number=false` almost immediately.

Otherwise the most extensive edits are to:
- use `-` instead of `*` as the top-level bullet-list marker, and
- move all link definitions to the bottom of the document.

Fixes #8470.