test(spanner): cater to the service vagaries on nested JSON
Reduce the "spanner_json_max_nesting_level" from 90 to 80, now that
the service is crashing on nesting depths somewhere in [82..90].
It appears that 80 will be the new documented limit.

And that means we temporarily expect `kAlreadyExists` when trying to
re-add a value with a nesting depth of 81.

See #7212, #8327, and #8689 for previous action in this area.