doc: leave a reference to the mdformat configuration file (#9733)
Mention the `.mdformat.toml` file in the "checkers" build script.
We need the configuration file to stop `mdformat` from searching
up outside the source tree, so we might as well use it to set
the configuration, instead of using command-line options.  This
breadcrumb will reinforce that convention.

Also add some hyphenation to a nearby message while I'm here.