test(storage): more consistent benchmark output and options
Now all benchmarks support `--labels`, which makes it easier to
concatenate and then analyze the results from multiple tests. The test
also capture GCE metadata by default, this is useful for analysis too.
I also changed the output from different benchmarks to start with the
timestamp of each experiment and then the labels.