test(storage): fix handling of optional ranges
I confused myself on how to interpret the default read and offset
ranges. The intent was to *not* set a `ReadRange()` option if the
command line arguments were not set. The effect was to set an empty
`ReadRange()` value, which produced an unexpected `Range:` header.
Fortunately, the value of the header had no effect, but really confused
me.

With this change the optional parameters are represented by
`absl::optional` values. Which makes it easier to treat them correctly.