ci: update builds in v1.42.x branch
Copy the build scripts from `main` so they work with the current Kokoro
and GCB configuration.

Revert the gRPC upgrades to use the latest gRPC version supporting
C++11

Hack the C++14 tests to test with C++11 instead. The script names are
confusing. With some forethought we may be able to avoid that problem.

Add all the Kokoro configuration files to support the current CI matrix.