feat(storage): easier mocks with `BucketMetadata`
Populating all the fields for `BucketMetadata` was somewhat difficult.
The class tried to prevent applications from setting fields that only
the server would set.  I thought about adding a `BucketMetadataBuilder`,
but that would duplicate a lot of code without preventing anything. In
addition, all the other libraries use protos that allow any field to be
set, and that does not seemt to create a lot of trouble or confusion.

I also stopped using the "composition-by-private-inheritance" base
because it does not save enough code.  I expect this class will go away
once I make similar changes to `ObjectMetadata`.