ci: compile all quickstarts on Windows (#9920)
For CMake: compile as many as are available with the `vcpkg`
version used in the builds.

For Bazel: compile as many as are available with the `google-cloud-cpp`
release used in the quickstart `WORKSPACE.bazel` files.

In both cases I created a hard-coded list of libraries that do not compile
on Windows.  I could not figure out how to query the libraries that
fail to compile on Windows from Bazel or CMake.