fix: spurious timeouts waiting for connection (#9927)
The helpers to wait for a `grpc::Channel` to be ready sometimes failed
because the channel was already connected. This was harmless, as this
was mostly used to continuously poll the connections.

The tests could flake under load.  Put some code to prevent this, and
increase the timeouts to make it even less likely.