cleanup(storage): minimize friends for `*PatchBuilder`
Use a single `friend` struct that needs to be forward declared.
The Google Style Guide frowns upon `struct` types with only static
functions. I think it is justified in this case because (a) it
simplifies all the `friend` declaratios, (b) it avoids exposing the
`nlohmann` stuff in our public headers.