fix(spanner): remove session from pool upon "not found" refresh failure
When the "SELECT 1" session-refresh query fails with "Session not found",
remove the bad session from the pool, if possible.  Previously we did
nothing because (1) it shouldn't happen, and (2) we have other mechanisms
in place to deal with bad sessions (which must remain).  But it is better
dealt with directly if we can.

Fixes #4026.