feat(storage): faster `InsertObject()` uploads (#9997)
With this change the client library will optimistically use a randomly generated string as a message boundary, but will not validate if the string appears in the payload.

The new behavior makes sense as (1) it saves about 2ms of client CPU time for 128MiB uploads, (2) if there is a collision this will be detected by the service via the upload checksums, and (3) the probability of a collision is small.

The current implementation has a population of more than $10^{100}$ strings to pick from. The probability of finding this string in a EB of data is less than $1 / 10^{95}$.  You could upload a EB of data per second for a billion years and still not find a collision.