test(storage): make one test more hermetic (#10019)
Even though `create_client_integration_test.cc` is an integration test,
there is no reason to use the default credentials with the emulator.
The default credentials typically try to use the GCE metadata service,
which may not exist in some CI environments. Or may take a long time to
fail.