feat(pubsub): implement blocking publisher
Some applications need a simple way to publish a small number of event
to several topics.  Or they need to publish with a low rate to many
topics. The `pubsub::BlockingPublisher` client is more convenient for
such applications than the existing `pubsub::Publisher`. The latter is
optimized for high-throughput clients publishing to a single topic.