fix(rest): return complete payloads for errors (#10051)
We need the full payload when returning completing a `HttpRequest`
with a non-OK status code. The payload may contain a full ErrorInfo, and
we want the application to have access to those.

`rest_internal::CurlImpl` was converting the HTTP status code unless it
was in a list of values ignored. Then `storage::RestClient` used a
per-method predicate to convert some values to errors.  It seems
simpler to *never* convert an HTTP status code to a `Status` in the
`CurlImpl` layer. After all, an HTTP request that completes with a
status code was able to send and receive the call.