test(spanner): restore creating an instance using the custom config
Undo the temporary measure that refrained from creating an `Instance` that
uses the custom `InstanceConfig`.  Apparently this was due to some rollout
skew in the backend, exacerbated by a freeze.

There remains some question about whether the `DeleteInstance()` should be
be followed by a `ListInstances()` polling loop, or the `DeleteInstanceConfig()`
should be retried while it returns "Cannot remove <name> since it is used",
but for now we proceed with neither.

Reverts #10033.  Fixes #10006.