cleanup(generator): improve test RPC names
The test RPCs had names inspired by their source service.  I found this
confusing when trying to determine if a new RPC for "streaming reads"
was needed, or what the right RPC for a "streaming read-write" test
would be.  This PR changes their names to `StreamingRead()`,
`StreamingWrite()`, and `StreamginReadWrite()`.  It also changes the
request and response messages to be shared across RPCs, and they are
simply named `Request` and `Response`.