feat(spanner): support for the PG.JSONB data type
Introduce `google::cloud::spanner::JsonB` for PostgreSQL databases.
The interface is the same as `Json` (just a strongly-typed string),
but `JsonB` values are stored in the database in a decomposed, binary
format, which affects performance: `JsonB` is slower to input, but
faster to process as it avoids reparsing.

Note: Support for arrays of `JSONB` will come later.