feat(spanner): add support for RowStream::RowsModified() (#10102)
Soon it will be possible for `Client::ExecuteQuery()` to execute a DML
statement (within a read/write transaction) that both modifies and
returns data.  So, if the result source behind the `RowStream` can
produce `ResultSetStats`, we can return the number of rows that were
modified.  Otherwise, for a normal query, return zero.