doc: add endpoint override snippets to generated libs (#10129)
Modify the landing pages for all the generated libraries to include a
snippet showing how to override the endpoint for the `*Client`.

In libraries where there is more than one `*Client` injecting a
snippet for each `*Client` looked too cluttered. The script injects the
first (in alphabetical order) snippet into the landing page. The
remaining snippets are referenced from that page, but are in
separate pages.