impl(common): helper functions to parse JSON configuration (#10274)
To support workload/workforce identity federation (aka BYOID,
aka external accounts) we will need to valid the JSON configuration.
This involves a lot of "is this field there and is it a string", or "if
the field is not there use this default, otherwise it must be a
string". These two functions save a lot of repetition in the validation
code.