fix(common): create default gRPC credentials only if needed
For backwards compatibility with the pre-Unified Credentials world we
were calling `grpc::GoogleDefaultCredentials()` if
`g::c::GrpcCredentialsOption` was not set. In many cases the option is
not set, but `g::c::UnifiedCredentialsOption` is. If that is the case,
the unified options are preferred. Calling
`grpc::GoogleDefaultCredentials()` produces spurious errors, and these
credentials won't be used.