ci: create build for external accounts integration
We need integration tests for external account support. The
existing CI systems run on Google infrastructure and therefore not
suitable for these tests. GitHub Actions are a bit underpowered, but
if we keep the number of tests in this builds small, and the caches are
warm, the builds are fast enough.