impl(common): a new decorator to cache credentials (#10319)
Multiple credential types are expensive to create. Fortunately, they
typically can be cached for about an hour. This PR adds a new member
function to `oauth2_internal::Credentials` that works better with
decorators for caching. The PR also includes a caching decorator,
but this is not used yet. Future PRs will change the code to use the new
function and decorator.