impl: simplify `oauth2_internal::ComputeEngineCredentials`
The implementation of `ComputeEngineCredentials` queries the metadata
service to support self-signed URLs in `google/cloud/storage`. This
query only needs to be done once, once the data is successfully
retrieved it is not going to change as it is immutable in the GCE
instance.

It also makes no sense to fail the token creation if the metadata query
fails.