impl(common): more consistent HTTP clients (#10349)
For external accounts we will need HTTP clients to exchange subject
tokens into access tokens.  Some (most?) subject token sources also need
HTTP clients. I think it will be simpler if the subject token sources
receive a `RestClient` factory when asked to fetch a subject token, as
opposed to binding a factory to the time when the source is created.

The downside is an extra (unused) parameter in the subject token sources
that do not use HTTP.