impl: propagate `g::c::Options` through credentials
Most credentials consume a `g::c::Options`.  In some places we were not
propagating these options from the `g::c::Make*Credentials()` call all
the way to the implementation class.

This change fixes that problem and starts the support for
`g::c::Options` in all credentials. The last bit would be useful to
enable logging when needed.