impl(common): decorator to log REST credentials
This will be used in the initialization of
`oauth2_internal::Credentials`.  I am planning to log any usage of
cached credentials, and also any refreshes.  Using
`ComputeEngineCredentials` as an example, and with all types implicitly
in the `oauth2_internal` namespace, the full typical stack would be:

```
LoggingCredentials(cached) -> CachedCredentials ->
     LoggingCredentials(refresh) -> ComputeEngineCredentials
```

The logging elements are optional. They will be inserted only if the
`g::c::TracingComponentsOption` includes `auth`.