ci: cleanup Kokoro macOS builds
Over the years we accumulated a number of steps to "fix" the Kokoro
macOS configuration.  We have updated the Kokoro base image a couple of
times, and the same problems are unlikely to still be there.  We also
kept going after `gcloud install` failed. The build breaks anyway, it is
better to stop the build early.