cleanup: shorter names for CMake targets
The CMake targets need to be unique across the whole project. That is
trivial for libraries, but tests often have conflicting names. A
`foo_test.cc` will typically result in a target name of `foo` regardless
of what directory it is in. Note that only the target name needs to be
globally unique, the output file (`foo_test`) is already in the same
directory as `foo_test.cc` and not likely to conflict.

In some places we used something inspired by the path (relative to the
project root or to `google/cloud`) to prefix the target name. In other
places we used something inspired by the associated library. I think it
is simpler to always use the path. And it also yields shorter target.