impl(oauth2): helper functions to compute `HMAC_SHA256()`
As part of the AWS external account support we need to generate a signed
request:

> Construct the AWS signed request (AWS Signature Version 4) using the
> `GetCallerIdentity`

(from)
https://google.aip.dev/auth/4117#determining-the-subject-token-in-aws

These signed requests are documented in:

https://docs.aws.amazon.com/general/latest/gr/create-signed-request.html

They require computing multiple `HMAC_SHA256()` values. The
implementation is simple if you already have a dependency on OpenSSL (as
we do). A longer description of these can be found in Wikipedia:

https://en.wikipedia.org/wiki/HMAC