cleanup(spanner): do not provide stub functions for unused RPCs
We do not provide a stub function for `Spanner.Read` because the client
implementation doesn't use it.  So, do the same for `Spanner.GetSession`
and `Spanner.ListSessions`, which are similarly unused.

This will make it clearer that `GetSession` (like `Read`) does not need
to be considered when adding routing-header support in the future.