impl(oauth2): building blocks for AWS external accounts (#10439)
As with the other external accounts, we need to build a subject token
for AWS external accounts. Building this subject token requires
obtaining a number of pieces of information. These can be found on
environment variables or via the VM's metadata service. The functions
to get them are complex enough that they deserve their own tests.