cleanup(storage): stop using XML for uploads and downloads (#10472)
We no longer have reason to believe that XML is faster than JSON for
uploads and/or downloads. The extra code carries significant complexity,
which is no longer justified. Removing this code is non-breaking, as the
API is the same, in fact, we already had support to "fallback" to JSON
when using features that XML does not support.