impl(storage): add the skeleton of the XML parser class
This parser is used for GCS MPU:
https://cloud.google.com/storage/docs/multipart-uploads

Also the `ParseXml` method is unimplemented, but it's too big
for a single PR, so I chose to split the PRs.

My intention for initializing regexes in the ctor is to only
compile the regex when creating the parser, although I'm not
sure if it really matters.