feat(generator): add some support for deprecated RPCs (#10573)
Do not generate APIs for deprecated RPCs unless their names appear in
an `emitted_rpcs` list, the reverse of the existing `omitted_rpcs`.

Furthermore, require that deprecated RPCs are listed in either of
`emitted_rpcs` or `omitted_rpcs`.  Use `emitted_rpcs` when the RPC
code has previously been generated (probably from before it was
deprecated), and so is needed for backwards compatibility.  Otherwise
use `omitted_rpcs`, typically for a new-generated service with
existing deprecated RPCs (which should be rare).

This will allow us to reliably omit generating APIs for deprecated
RPCs that we have never supported.

Part of #8486.