feat(generator): more support for deprecated RPC overloads (#10583)
Require that all "method_signature" overloads that use deprecated
fields appear in either the `emitted_rpcs` or `omitted_rpcs` lists.
This allows us to distinguish between RPC overloads using a newly-
deprecated field, which we want to keep generating for backwards
compatibility, and RPC overloads using deprecated fields in newly-
generated services, which we want to ignore.

Add `omitted_rpcs` entries for RPC overloads we currently ignore.

A subtle change is that we now compute the full "signature uid" for
all overloads, not just the un-deprecated ones.  This is a feature.
Without it, we ran the risk of choosing a new "first match" from a
clashing set should a field be newly deprecated, and therefore
silently changing the meaning of existing programs.

Add some simple support for allowing commas in the elements of
`emitted_rpcs` and `omitted_rpcs`.

Part of #8486.