cleanup(storage): a handful of concluding cleanups for XmlParser
[Reviewers: Yes, this is somewhat of a catchall, but it is convenient
to gather all these together to finalize this work for now.]

Reformulate the `XmlParser` as a true state machine, where there
are no branches on the next input char that are independent of the
current state.  This means the current state can be represented
simply by the handler function, without maintaining a parallel
`enum`.

Correct a couple of transitions to the "base" state instead of the
"end tag" state.  This should not affect a successful parse, but
will improve the message given for errors after the transition.

There is no need to store `root_` and `current_parent_` separately
when they can just be the bottom and top of `node_stack_`.

Accumulate the start/end tag names and the text content as views
into the XML text, to eliminate copying into separate strings.

Check, when the input has been consumed, that we're not in the
middle of an element.

Count the root node towards the `XmlParserMaxNodeDepth` limit.

Use `ASSERT_STATUS_OK(s)` rather than `ASSERT_TRUE(s.ok())` so that
we get informative failures.

Use `ASSERT_STATUS_OK(s_or)` instead of `EXPECT_STATUS_OK(s_or)`
when the next thing we would do is look at the `StatusOr`s value.

Rename all the `XmlNodeTest.Case` test cases to just `XmlNode.Case`.