ci: deflake checks and check-api builds (#10754)
These builds used `bazel query` (or `bazelisk query`) to discover the
list of GA libraries. Starting Bazel requires downloading dependencies,
which can fail in the CI system. In normal Bazel builds we minimize
these failures with some  `bazel fetch` before the build, but now that
the list of GA libraries is in CMake there is a simpler and less flaky
way to discover them.