cleanup(bigtable): remove unused CMake option (#10818)
Remove some code and CMake option used to force a static analysis error.
We no longer have any builds with this static analysis, and when we did,
the flag was used to troubleshoot problems with the build not failing
when it should. The latter is better done with a local change rather
than polluting the options namespace.