impl(spanner): add TransactionContext.route_to_leader
RPCs within the transaction will then be able to inspect the context
to determine if their requests should be routed to the region leader.

Mostly this is just YES for read-write transactions and NO for readonly
transactions.  Partitioned-DML transactions are a slight complication,
which will be dealt with presently.

Query and read partitions currently say NO, pending some updates as to
how they are serialized/deserialized.