cleaup(spanner): rename the "update-database" snippet tag
We made a poor choice in naming the (hand-written) `DatabaseAdminClient`
interface to `rpc UpdateDatabaseDdl` as just `UpdateDatabase()`.  Soon
there actually will be a database-admin `UpdateDatabase` RPC, which
fortunately we'll only support in the generated API, but it will want to
use the "update-database" tag.

So, use "add-column" for the old snippet tag instead.  It is also a
closer match for the corresponding doc tag, "spanner_add_column".