impl(spanner): support route_to_leader in query/read partitions (#10918)
Add "route to leader" support to `QueryPartition` and `ReadPartition`
so that RPCs in the eventual partitioned `ExecuteQuery()` and `Read()`
calls will be routed according to the transaction used in the original
`PartitionQuery()` and `PartitionRead()` calls (respectively).

Note that we support (de)serialization of the `QueryPartition` and
`ReadPartition` values so that they might be transferred to another
address space, and that this is done using `ExecuteSqlRequest` and
`ReadRequest` protobuf messages, which do not include route-to-leader
fields.  The best (only?) way to handle this while maintaining
backwards compatibility is to add the route-to-leader value to those
messages using a protobuf unknown field.