fix(generator): ensure warnings and errors are sent to std::clog
If the generator is to use `GCP_LOG()` to report problems, then we
must ensure that logging messages end up on the "stderr" output stream.
The easiest way to do that is enable the `std::clog` backend on the
`LogSink` (flush/tie aside), which we were already attempting.

Unfortunately the `EnableStdClog()` call was doing nothing because a
default backend was already in place.  So, we must remove that backend
first, and the easiest was to do that is to call `DisableStdClog()`.

Also lower the default minimum log severity to NOTICE, which is the
first level that is an indication of a problem.

And finally add a `--log_level` flag so that the minimum can be
lowered all the way down to the compile-time limit.  Parsing this
flag required exposing `ParseSeverity()`.